+++ packages/terra-base/tests/nightwatch/DefaultBase.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
-
 import Base from '../../lib/Base';
-
 
 const BaseDefault = () => (
   <Base>