+++ packages/terra-site/CHANGELOG.md
@@ -5,6 +5,8 @@ Unreleased
 ----------
 ### Added
 * Add terra-form-checkbox components
+* Add terra-form-field components
+* Add terra-mock-theme styles
 
 1.17.0 - (November 7, 2017)
 ------------------